requirements.txt
pandas
pyyaml
python-dotenv
google-generativeai
pytest
python-dotenv